+++ src/Cirreum.Identity.EntraExternalId/EntraExternalId/EntraExternalIdJsonContext.cs
@@ -5,5 +5,7 @@ using Cirreum.Identity.EntraExternalId.Models;
 
 [JsonSerializable(typeof(EntraClaimsRequest))]
 [JsonSerializable(typeof(EntraClaimsResponse))]
+[JsonSerializable(typeof(string))]
+[JsonSerializable(typeof(string[]))]
 internal sealed partial class EntraExternalIdJsonContext : JsonSerializerContext {
 }